# Show nanoseconds in strace log