This can be useful when matching the strace log with the shadow log, or for seeing where syscall latency is applied.

Before:

```
00:00:01.000 [tid 1000] socket(Inet, 1, 0) = 3
00:00:01.000 [tid 1000] socket(Inet, 1, 0) = 4
00:00:01.000 [tid 1000] bind(3, 0x7fffffffdc88, 16) = 0
00:00:01.000 [tid 1000] getsockname(3, 0x7fffffffdc88, 16 (0x7fffffffdcf0)) = 0
00:00:01.000 [tid 1000] listen(3, 10) = 0
00:00:01.000 [tid 1000] connect(4, 0x7fffffffddec, 16) = 0
00:00:01.002 [tid 1000] nanosleep(...) = 0
00:00:01.002 [tid 1000] connect(4, 0x7fffffffe14c, 16) = -106 (EISCONN: Transport endpoint is already connected)
00:00:01.002 [tid 1000] close(4) = 0
00:00:01.002 [tid 1000] close(3) = 0
```

After:

```
00:00:01.000097028 [tid 1000] socket(Inet, 1, 0) = 3
00:00:01.000099028 [tid 1000] socket(Inet, 1, 0) = 4
00:00:01.000099028 [tid 1000] bind(3, 0x7fffffffdc88, 16) = 0
00:00:01.000101028 [tid 1000] getsockname(3, 0x7fffffffdc88, 16 (0x7fffffffdcf0)) = 0
00:00:01.000101028 [tid 1000] listen(3, 10) = 0
00:00:01.000103030 [tid 1000] connect(4, 0x7fffffffddec, 16) = 0
00:00:01.002103030 [tid 1000] nanosleep(...) = 0
00:00:01.002103030 [tid 1000] connect(4, 0x7fffffffe14c, 16) = -106 (EISCONN: Transport endpoint is already connected)
00:00:01.002105030 [tid 1000] close(4) = 0
00:00:01.002105030 [tid 1000] close(3) = 0
```